backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.7.6.dev0
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
apturl==0.5.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-storage-common==2.1.0
Babel==2.8.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.44
botocore==1.34.44
Brlapi==0.8.3
cachetools==5.4.0
capstone==5.0.1
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.6.7
dbus-python==1.2.18
debtcollector==3.0.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
dogpile.cache==1.3.3
ecdsa==0.18.0
elementpath==4.4.0
email_validator==2.2.0
enum-compat==0.0.3
eventlet==0.36.1
exceptiongroup==1.2.2
fastapi==0.112.2
fastapi-cli==0.0.5
fastapi-limiter==0.1.6
fasteners==0.19
filelock==3.6.0
flake8==6.1.0
Flask==3.0.3
Flask-RESTful==0.3.10
frozenlist==1.4.1
fsspec==2024.6.1
fusepy==3.0.1
future==0.18.3
futurist==3.0.0
greenlet==3.0.1
gyp==0.1
h11==0.14.0
hashID==3.1.4
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib-metadata==4.6.4
iniconfig==2.0.0
intuit-oauth==1.2.4
iso8601==2.1.0
isodate==0.6.1
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
keystone==25.0.0
keystone-engine==0.9.2
keystoneauth1==5.7.0
keystonemiddleware==10.7.0
kombu==5.3.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama-cloud==0.0.15
llama-index==0.11.4
llama-index-agent-openai==0.3.0
llama-index-cli==0.3.0
llama-index-core==0.11.4
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.4
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-ollama==0.3.1
llama-index-llms-openai==0.2.1
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.0
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.1
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mccabe==0.7.0
mdurl==0.1.2
minijinja==2.2.0
mock==5.1.0
more-itertools==8.10.0
mpmath==1.3.0
msal==1.26.0
msal-extensions==1.1.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netaddr==1.3.0
netifaces==0.11.0
networkx==3.3
nltk==3.9.1
nodeenv==1.8.0
nulltype==2.3.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
ollama==0.3.2
openai==1.43.0
os-service-types==1.7.0
oslo.cache==3.8.0
oslo.concurrency==6.0.0
oslo.config==9.5.0
oslo.context==5.5.0
oslo.db==16.0.0
oslo.i18n==6.3.0
oslo.log==6.1.1
oslo.messaging==14.8.1
oslo.metrics==0.8.0
oslo.middleware==6.1.0
oslo.policy==4.4.0
oslo.serialization==5.4.0
oslo.service==3.5.0
oslo.upgradecheck==2.3.0
oslo.utils==7.2.0
osprofiler==4.1.0
packaging==23.2
pandas==2.2.2
passlib==1.7.4
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.0.0
pexpect==4.8.0
Pillow==9.0.1
plaid==0.1.7
plaid-python==17.0.0
platformdirs==2.5.1
pluggy==1.3.0
portalocker==2.8.2
prettytable==3.10.2
prometheus_client==0.20.0
protobuf==3.12.4
proton-core==0.2.0
proton-keyring-linux==0.0.2
proton-keyring-linux-secretservice==0.0.2
proton-vpn-api-core==0.32.2
proton-vpn-gtk-app==4.4.4
proton-vpn-killswitch-network-manager==0.5.4
proton-vpn-killswitch-network-manager-wireguard==0.1.4
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.5.2
proton-vpn-network-manager-openvpn==0.1.0
proton-vpn-network-manager-wireguard==0.4.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.5.1
pycadf==3.1.1
pycairo==1.20.1
pycodestyle==2.11.1
pycryptodome==3.20.0
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.1.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.9.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==4.3.1
pyRFC3339==1.1
pysaml2==7.5.0
pyserial==3.5
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
pytest==7.4.3
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-jose==3.3.0
python-keystoneclient==5.4.0
python-multipart==0.0.9
pytz==2022.1
pyusb==1.2.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==6.0.2
redis==5.1.0b7
referencing==0.35.1
regex==2024.7.24
reportlab==3.6.8
repoze.lru==0.7
requests==2.32.3
requests-oauthlib==1.3.1
rfc3986==2.0.0
rich==13.8.0
Routes==2.5.1
rpds-py==0.19.1
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
scrypt==0.8.24
SecretStorage==3.3.1
sentence-transformers==3.0.1
sentry-sdk==1.4.3
shellingham==1.5.4
shiboken6==6.7.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.23
sqlalchemy-cockroachdb==2.0.1
ssh-import-id==5.11
starlette==0.38.2
statsd==4.0.1
stevedore==5.2.0
striprtf==0.0.26
sympy==1.13.2
systemd-python==234
tabulate==0.8.9
tenacity==8.5.0
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.3
torch==2.4.1
tqdm==4.66.4
transformers==4.44.2
triton==3.0.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unicorn==2.0.1.post1
UNKNOWN @ file:///home/runmy/mtkclient
urllib3==1.26.5
uvicorn==0.30.6
uvloop==0.20.0
vine==5.1.0
virtualenv==20.13.0+ds
wadllib==1.3.6
watchfiles==0.24.0
wcwidth==0.2.13
WebOb==1.8.7
websockets==13.0.1
Werkzeug==3.0.3
wrapt==1.16.0
xdg==5
xkit==0.0.0
xmlschema==2.5.1
yappi==1.6.0
yarl==1.9.4
youtube-dl==2021.12.17
zipp==1.0.0
